+++ src/USBHostSerial.cpp
@@ -14,7 +14,7 @@ SPDX-License-Identifier: CC0-1.0
 
 using namespace esp_usb;
 
-USBHostSerial::USBHostSerial()
+USBHostSerial::USBHostSerial(uint16_t vid, uint16_t pid)
 : _host_config{}
 , _line_coding{}
 , _tx_buf_mem{}
@@ -24,7 +24,9 @@ USBHostSerial::USBHostSerial()
 , _rx_buf_handle(nullptr)
 , _rx_buf_data{}
 , _setupDone(false)
-, _connected(false)
+, _fallback(false)
+, _vid(vid)
+, _pid(pid)
 , _device_disconnected_sem(nullptr)
 , _usb_lib_task_handle(nullptr)
 , _logger(nullptr) {
@@ -43,7 +45,11 @@ USBHostSerial::~USBHostSerial() {
 }
 
 USBHostSerial::operator bool() const {
-  return _connected;
+  if (xSemaphoreTake(_device_disconnected_sem, 0) == pdTRUE) {
+    xSemaphoreGive(_device_disconnected_sem);
+    return false;
+  }
+  return true;
 }
 
 bool USBHostSerial::begin(int baud, int stopbits, int parity, int databits) {
@@ -71,10 +77,8 @@ void USBHostSerial::end() {
 
 std::size_t USBHostSerial::write(uint8_t data) {
   if (xRingbufferSend(_tx_buf_handle, &data, 1, pdMS_TO_TICKS(1)) == pdTRUE) {
-    _log("USB buf write: 1");
     return 1;
   }
-  _log("USB buf write: 0");
   return 0;
 }
 
@@ -88,9 +92,6 @@ std::size_t USBHostSerial::write(const uint8_t *data, std::size_t len) {
     _log(buf);
   } else {
     if (xRingbufferSend(_tx_buf_handle, data, len, pdMS_TO_TICKS(1)) == pdTRUE) {
-      char buf[30];
-      snprintf(buf, 30, "USB buf write: %u", len);
-      _log(buf);
       return len;
     }
   }
@@ -111,7 +112,6 @@ uint8_t USBHostSerial::read() {
     retVal = *reinterpret_cast<uint8_t*>(ret);
     vRingbufferReturnItem(_rx_buf_handle, ret);
   }
-  _log("USB buf read: 1");
   return retVal;
 }
 
@@ -128,9 +128,6 @@ std::size_t USBHostSerial::read(uint8_t *dest, std::size_t size) {
       break;
     }
   }
-  char buf[30];
-  snprintf(buf, 30, "USB buf read: %u", retVal);
-  _log(buf);
   return retVal;
 }
 
@@ -141,7 +138,7 @@ void USBHostSerial::setLogger(USBHostSerialLoggerFunc logger) {
 void USBHostSerial::_setup() {
   _device_disconnected_sem = xSemaphoreCreateBinary();
   assert(_device_disconnected_sem);
-  xSemaphoreGive(_device_disconnected_sem);
+  xSemaphoreGive(_device_disconnected_sem);  // make available for first use
 
   // Install USB Host driver. Should only be called once in entire application
   _host_config.skip_phy_setup = false;
@@ -169,16 +166,12 @@ bool USBHostSerial::_handle_rx(const uint8_t *data, size_t data_len, void *arg) 
     // log overflow warning
     static_cast<USBHostSerial*>(arg)->_log("USB rx buf overflow");
   }
-  char buf[30];
-  snprintf(buf, 30, "USB rx: %u", lenReceived);
-  static_cast<USBHostSerial*>(arg)->_log(buf);
   return true;
 }
 
 void USBHostSerial::_handle_event(const cdc_acm_host_dev_event_data_t *event, void *user_ctx) {
   if (event->type == CDC_ACM_HOST_DEVICE_DISCONNECTED) {
     xSemaphoreGive(static_cast<USBHostSerial*>(user_ctx)->_device_disconnected_sem);
-    static_cast<USBHostSerial*>(user_ctx)->_connected = false;
   }
 }
 
@@ -194,6 +187,7 @@ void USBHostSerial::_usb_lib_task(void *arg) {
 
 void USBHostSerial::_USBHostSerial_task(void *arg) {
   USBHostSerial* thisInstance = static_cast<USBHostSerial*>(arg);
+  esp_err_t err = ESP_OK;  // reusable
   while (1) {
     // try to open USB VCP device
     const cdc_acm_host_device_config_t dev_config = {
@@ -204,41 +198,59 @@ void USBHostSerial::_USBHostSerial_task(void *arg) {
       .data_cb = _handle_rx,
       .user_arg = thisInstance,
     };
-    vTaskDelay(pdMS_TO_TICKS(10));
+    cdc_acm_dev_hdl_t cdc_dev = NULL;
     auto vcp = std::unique_ptr<CdcAcmDevice>(VCP::open(&dev_config));
-    vTaskDelay(pdMS_TO_TICKS(10));
     if (vcp == nullptr) {
-      continue;
-    } else {
+      // try to fallback to CDC
+      err = cdc_acm_host_open(thisInstance->_vid, thisInstance->_pid, 0, &dev_config, &cdc_dev);
+      if (err != ESP_OK) {
+        continue;
+      }
+      thisInstance->_fallback = true;
       thisInstance->_log("USB CDC device opened");
+    } else {
+      thisInstance->_fallback = false;
+      thisInstance->_log("USB VCP device opened");
     }
 
-    // Mark connected and configure
-    thisInstance->_connected = true;
+    // mark connected
     xSemaphoreTake(thisInstance->_device_disconnected_sem, portMAX_DELAY);
-    if (vcp->line_coding_set(&(thisInstance->_line_coding)) == ESP_OK) {
+
+    // set line coding
+    err = ESP_OK;
+    if (thisInstance->_fallback) {
+      err = cdc_acm_host_line_coding_get(cdc_dev, &(thisInstance->_line_coding));
+    } else {
+      err = vcp->line_coding_set(&(thisInstance->_line_coding));
+    }
+    if (err == ESP_OK) {
       thisInstance->_log("USB line coding set");
     } else {
       thisInstance->_log("USB line coding error");
       continue;
     }
 
+    // all set, enter loop to start sending
     while (1) {
       // check if still connected
       if (xSemaphoreTake(thisInstance->_device_disconnected_sem, 0) == pdTRUE) {
-        thisInstance->_connected = false;
         break;
       }
 
       // check for data to send
       std::size_t pxItemSize = 0;
-      uint8_t *data = (uint8_t*)xRingbufferReceiveUpTo(thisInstance->_tx_buf_handle, &pxItemSize, pdMS_TO_TICKS(10), USBHOSTSERIAL_BUFFERSIZE);
-      if (data && thisInstance->_connected) {
-        char buf[30];
-        snprintf(buf, 30, "USB tx: %c...(%u)", data[0], pxItemSize);
-        thisInstance->_log(buf);
-        ESP_ERROR_CHECK(vcp->tx_blocking(data, pxItemSize, 1000));
-        vRingbufferReturnItem(thisInstance->_tx_buf_handle, (void*)data);
+      void *data = xRingbufferReceiveUpTo(thisInstance->_tx_buf_handle, &pxItemSize, pdMS_TO_TICKS(10), USBHOSTSERIAL_BUFFERSIZE);
+      if (data) {
+        if (thisInstance->_fallback) {
+          err = cdc_acm_host_data_tx_blocking(cdc_dev, (uint8_t*)data, pxItemSize, 1000);
+        } else {
+          err = vcp->tx_blocking((uint8_t*)data, pxItemSize, 1000);
+        }
+        if (err == ESP_OK) {
+          vRingbufferReturnItem(thisInstance->_tx_buf_handle, data);
+        } else {
+          thisInstance->_log("Error writing to USB");
+        }
       }
       taskYIELD();
     }