+++ libraries/usb_host_cdc_acm/CHANGELOG.md
@@ -1,3 +1,27 @@
+## 2.1.1
+
+- Added support for ESP32-H4
+
+## 2.1.0
+
+- Added option to implement custom CDC-ACM like devices with C API
+
+## 2.0.6
+
+- Fixed device opening for devices with CDC class defined in Device descriptor https://github.com/espressif/esp-usb/pull/89
+- Fixed driver installation from a task with priority >= CDC driver priority
+
+## 2.0.5
+
+- Fixed CDC descriptor parsing logic, when some CDC devices could not be opened
+- Added an option to open a CDC device with any VID/PID
+- Fixed crash on ESP32-P4 if Receive buffer append function (introduced in v2.0.0) was used; this function does not work on P4 yet
+
+## 2.0.4
+
+- Fixed Control transfer allocation size for too small EP0 Max Packet Size (https://github.com/espressif/esp-idf/issues/14345)
+- Merged `open()` and `open_vendor_specific()` functions. All types of CDC devices are now opened with `cdc_acm_host_open()`, CDC compliance is detected automatically
+
 ## 2.0.3
 
 - Added `cdc_acm_host_cdc_desc_get()` function that enables users to get CDC functional descriptors of the device